+++ src/collective/fingerpointing/tests/test_upgrades.py
@@ -120,3 +120,41 @@ class To3TestCase(BaseUpgradeTestCase):
         self._do_upgrade(step)
         self.assertEqual(user_actions.keys()[-1], 'logout')
         self.assertEqual(user_actions.keys()[-2], 'audit-log')
+
+
+class To4TestCase(BaseUpgradeTestCase):
+
+    from_ = '3'
+    to_ = '4'
+
+    def test_profile_version(self):
+        version = self.setup.getLastVersionForProfile(self.profile_id)[0]
+        self.assertEqual(version, self.from_)
+
+    def test_registered_steps(self):
+        steps = len(self.setup.listUpgrades(self.profile_id)[0])
+        self.assertEqual(steps, 1)
+
+    def test_update_user_actions(self):
+        # check if the upgrade step is registered
+        title = u'Add new field audit_workflow to configlet'
+        step = self._get_upgrade_step_by_title(title)
+        assert step is not None
+
+        from collective.fingerpointing.interfaces import IFingerPointingSettings
+        from plone.registry.interfaces import IRegistry
+        from zope.component import getUtility
+        registry = getUtility(IRegistry)
+
+        # simulate state on previous version
+        record = IFingerPointingSettings.__identifier__ + '.audit_workflow'
+        del registry.records[record]
+
+        with self.assertRaises(KeyError):
+            registry.forInterface(IFingerPointingSettings)
+
+        # execute upgrade step and verify changes were applied
+        self._do_upgrade(step)
+        settings = registry.forInterface(IFingerPointingSettings)
+        self.assertTrue(hasattr(settings, 'audit_workflow'))
+        self.assertEqual(settings.audit_workflow, True)